feat(model): add HTTP types to allow custom http instances
