+++ src/types/http.ts
@@ -1,0 +1,33 @@
+type Method =
+  | 'get'
+  | 'GET'
+  | 'delete'
+  | 'DELETE'
+  | 'head'
+  | 'HEAD'
+  | 'options'
+  | 'OPTIONS'
+  | 'post'
+  | 'POST'
+  | 'put'
+  | 'PUT'
+  | 'patch'
+  | 'PATCH'
+  | 'purge'
+  | 'PURGE'
+  | 'link'
+  | 'LINK'
+  | 'unlink'
+  | 'UNLINK'
+
+export type HTTPRequestConfig = {
+  method?: Method
+  url?: string
+  data?: Record<string, any> | any
+}
+
+export type HTTPResponse<T> = {
+  data: T
+}
+
+export type HTTPPromise<T = any> = Promise<HTTPResponse<T>>
